Refactor create_image tasks
This patch changes the ISO image mount tasks to use
`ansible.builtin.command` instead `ansible.posix.mount` to avoid adding
a depenency to the `ansbile.posix` collection.

The names of the API and Worker images can now be provided with role
variables.  The default values are the previously hardcoded
`openstack-barbican-api` and `openstack-barbican-worker`.