Fix inconsistency in copying certificates
The `ansible.builtin.copy` task behaves differently when copying
directories depending on whether a trailing slash is used or not.

This results in the secret sometimes not containing the certificates
when `proteccio_crt_src` did not end in a `'/'`.

This patch changes the copy task to always use a trailing slash.