Use get_url to copy certificates
This patch changes the task to copy certificate files to use
`ansible.builtin.get_url` to fetch the server certificate and the client
cert and key individually instead of copying the contents of a directory
with `ansible.builtin.copy`.

This requires a couple more role variables, but it gives us flexibility
to store the files in a remote location.