Add meta/main.yml to make role installable via ansible-galaxy
The role currently cannot be installed using ansible-galaxy due to
the missing meta/main.yml file, which is required for Ansible Galaxy
to recognize a directory as a valid role.

This commit adds the meta/main.yml file with:
- Role metadata (name, author, description, license)
- Minimum Ansible version requirement (2.9)
- Supported platforms (RHEL 8 and 9)
- Galaxy tags for discoverability
- Empty dependencies list

Error reproduced:
$ ansible-galaxy role install git+https://github.com/openstack-k8s-operators/ansible-role-rhoso-proteccio-hsm.git,main
[WARNING]: - ansible-role-rhoso-proteccio-hsm was NOT installed
successfully: this role does not appear to have a meta/main.yml file.

Signed-off-by: Mauricio Harley <mharley@redhat.com>